fix VulnManagement
