+++ web/src/pages/index.js
@@ -9,7 +9,7 @@ import { SignUp } from "./SignUp/SignUpPage";
 import { Status } from "./Status/StatusPage";
 import { Tag } from "./Tag/TagPage";
 import { TopicDetail } from "./TopicDetail/TopicDetailPage";
-import { TopicManagement } from "./TopicManagement/TopicManagementPage";
+import { VulnManagement } from "./VulnManagement/VulnManagementPage";
 import { Vulnerability } from "./Vulnerability/VulnerabilityPage";
 
 export {
@@ -24,6 +24,6 @@ export {
   Status,
   Tag,
   TopicDetail,
-  TopicManagement,
+  VulnManagement,
   Vulnerability,
 };
